requirements.txt
boto3==1.28.78
dj-database-url==0.5.0
Django==4.2
gunicorn==20.1.0
pycodestyle==2.7.0
python-decouple==3.4
whitenoise==5.3.0
pillow
# mysqlclient==2.2.4
num2words==0.5.13
django-storages==1.14.2
psycopg2-binary==2.9.9
reportlab==4.1.0
qrcode==7.4.2
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
django-environ==0.10.0
XlsxWriter==3.2.0

